Renamed filename for SyncSocketProvider
